modularized modal
