+++ src/routes/canvas/+page.server.ts
@@ -1,22 +1,22 @@
 import { error } from "@sveltejs/kit";
 import type { PageServerLoad } from "./$types";
 
-export const load: PageServerLoad = async ({url, locals: { supabase, safeGetSession } }) => {
+export const load: PageServerLoad = async ({ url, locals: { supabase, safeGetSession } }) => {
 	const { session } = await safeGetSession();
 	void session;
 
 	// TODO: Anonymous users can actually view the canvas too (as long as they are within the range)
 
 	// get backup code parameter (ex: /canvas?c=qwer-tyui-opas)
-	const backup_code: string | null = url.searchParams.get("c");
+	const backupCode: string | null = url.searchParams.get("c");
 
 	// parameter null-checking (ex: /canvas without c parameter)
-	if (backup_code === null) {
+	if (backupCode === null) {
 		error(404);
 	}
-	
+
 	// canvas existence check and archive check
-	const channelName = backup_code;
+	const channelName = backupCode;
 	const { data } = await supabase
 		.from("canvas")
 		.select("drawing, id, backup_code, is_archived")
