+++ static/stylesheets/guideline.css
@@ -1,1 +1,1 @@
-@media screen and (min-width: 621px){.only-phone{display:none !important}}@media screen and (max-width: 621px){.only-desktop{display:none !important}}.item-frame{border:1.5px solid #362121;border-radius:8px;border-bottom-width:4px}.no-drag{-webkit-user-drag:none;-moz-user-select:none;user-select:none;-webkit-user-select:none;-ms-user-select:none}.glass-heavy,.glass-light{background-color:rgba(236,120,70,.7);backdrop-filter:blur(20px);-webkit-backdrop-filter:blur(20px)}.glass-light{background-color:rgba(236,120,70,.4)}button{padding:11px 31.5px;height:50px;border:none;display:flex;justify-content:center;align-items:center;background-color:#362121;color:#ec7846;border-radius:8px;font-family:"Outfit",system-ui,Helvetica,sans-serif;font-size:16px;line-height:22px;font-weight:400;cursor:pointer;-webkit-tap-highlight-color:rgba(0,0,0,0);-webkit-user-drag:none;-moz-user-select:none;user-select:none;-webkit-user-select:none;-ms-user-select:none;transition:300ms cubic-bezier(0.3, 0.1, 0.1, 1);transition-property:opacity,transform,color,background-color,border-color}@media(hover: hover){button:hover{opacity:.7}}button:active{opacity:1;transform:scale(0.925)}button:disabled{opacity:.3 !important;pointer-events:none !important}button.solid{background:#362121}button.small{padding:11px 20px;height:42px;font-size:14px}button.outline,button.outline.small{background:#ec7846;color:#362121;border:1.5px solid #362121;border-radius:8px;border-bottom-width:4px;margin:0px;padding:11px 30px}@media(hover: hover){button.outline:hover{opacity:.5}}button.outline:active{opacity:1;margin-top:1.5px;height:48.5px;border-bottom-width:1.5px;transform:translateY(1.5px)}button.outline.small{padding:11px 20px;height:42px;font-size:14px}button.outline.small:active{height:40.5px;border-bottom-width:1.5px;transform:translateY(1.5px)}button.none{all:unset;transition:200ms cubic-bezier(0.3, 0.1, 0.1, 1);transition-property:opacity,transform;-webkit-tap-highlight-color:rgba(0,0,0,0)}@media(hover: hover){button.none:hover{opacity:.5}}button.none:active{opacity:1;transform:scale(0.8)}input,textarea{background-color:#ec7846;border:1.5px solid #362121;border-bottom-width:4px;border-radius:8px;padding:0px 20px;height:60px;font-family:"Outfit",system-ui,Helvetica,sans-serif;font-size:16px;line-height:22px;color:#5a3030;font-weight:400;-webkit-tap-highlight-color:rgba(0,0,0,0);outline:none;-webkit-user-drag:none;-moz-user-select:none;user-select:none;-webkit-user-select:none;-ms-user-select:none;transition:150ms cubic-bezier(0.37, 0.13, 0.33, 1);transition-property:color,border-color,opacity}input::-moz-placeholder, textarea::-moz-placeholder{color:#964b3c}input::placeholder,textarea::placeholder{color:#964b3c}input.invalid,textarea.invalid{color:#b11013;border-color:#b11013}input.invalid::-moz-placeholder, textarea.invalid::-moz-placeholder{color:#b11013}input.invalid::placeholder,textarea.invalid::placeholder{color:#b11013}textarea{padding:20px;height:auto;resize:none;transition:150ms cubic-bezier(0.37, 0.13, 0.33, 1);transition-property:color,border-color,opacity;-ms-overflow-style:none;scrollbar-width:none}textarea::-webkit-scrollbar{display:none}textarea.invalid{color:#b11013;border-color:#b11013}textarea.invalid::-moz-placeholder{color:#b11013}textarea.invalid::placeholder{color:#b11013}a{-webkit-tap-highlight-color:rgba(0,0,0,0)}a:active{opacity:.8}a.wrapper{text-decoration:none}@keyframes spin{to{transform:rotate(360deg)}}.animate-spin{animation:spin 1s linear infinite}
+@media screen and (min-width: 621px){.only-phone{display:none !important}}@media screen and (max-width: 621px){.only-desktop{display:none !important}}.item-frame{border:1.5px solid #362121;border-radius:8px;border-bottom-width:4px}.no-drag{-webkit-user-drag:none;-moz-user-select:none;user-select:none;-webkit-user-select:none;-ms-user-select:none}.glass-heavy,.glass-light{background-color:rgba(236,120,70,.7);backdrop-filter:blur(20px);-webkit-backdrop-filter:blur(20px)}.glass-light{background-color:rgba(236,120,70,.4)}button{padding:11px 31.5px;height:50px;border:none;display:flex;justify-content:center;align-items:center;background-color:#362121;color:#ec7846;border-radius:8px;font-family:"Outfit",system-ui,Helvetica,sans-serif;font-size:16px;line-height:22px;font-weight:400;cursor:pointer;-webkit-tap-highlight-color:rgba(0,0,0,0);-webkit-user-drag:none;-moz-user-select:none;user-select:none;-webkit-user-select:none;-ms-user-select:none;transition:300ms cubic-bezier(0.3, 0.1, 0.1, 1);transition-property:opacity,transform,color,background-color,border-color}@media(hover: hover){button:hover{opacity:.7}}button:active{opacity:1;transform:scale(0.925)}button:disabled{opacity:.3 !important;pointer-events:none !important}button.solid{background:#362121}button.small{padding:11px 20px;height:42px;font-size:14px}button.outline,button.outline.small{background:#ec7846;color:#362121;border:1.5px solid #362121;border-radius:8px;border-bottom-width:4px;margin:0px;padding:11px 30px}@media(hover: hover){button.outline:hover{opacity:.5}}button.outline:active{opacity:1;margin-top:1.5px;height:48.5px;border-bottom-width:1.5px;transform:translateY(1.5px)}button.outline.small{padding:11px 20px;height:42px;font-size:14px}button.outline.small:active{height:40.5px;border-bottom-width:1.5px;transform:translateY(1.5px)}button.none{all:unset;transition:200ms cubic-bezier(0.3, 0.1, 0.1, 1);transition-property:opacity,transform;-webkit-tap-highlight-color:rgba(0,0,0,0)}@media(hover: hover){button.none:hover{opacity:.5}}button.none:active{opacity:1;transform:scale(0.8)}button.hidden{all:unset;pointer-events:none;display:none}input,textarea{background-color:#ec7846;border:1.5px solid #362121;border-bottom-width:4px;border-radius:8px;padding:0px 20px;height:60px;font-family:"Outfit",system-ui,Helvetica,sans-serif;font-size:16px;line-height:22px;color:#5a3030;font-weight:400;-webkit-tap-highlight-color:rgba(0,0,0,0);outline:none;-webkit-user-drag:none;-moz-user-select:none;user-select:none;-webkit-user-select:none;-ms-user-select:none;transition:150ms cubic-bezier(0.37, 0.13, 0.33, 1);transition-property:color,border-color,opacity}input::-moz-placeholder, textarea::-moz-placeholder{color:#964b3c}input::placeholder,textarea::placeholder{color:#964b3c}input.invalid,textarea.invalid{color:#b11013;border-color:#b11013}input.invalid::-moz-placeholder, textarea.invalid::-moz-placeholder{color:#b11013}input.invalid::placeholder,textarea.invalid::placeholder{color:#b11013}textarea{padding:20px;height:auto;resize:none;transition:150ms cubic-bezier(0.37, 0.13, 0.33, 1);transition-property:color,border-color,opacity;-ms-overflow-style:none;scrollbar-width:none}textarea::-webkit-scrollbar{display:none}textarea.invalid{color:#b11013;border-color:#b11013}textarea.invalid::-moz-placeholder{color:#b11013}textarea.invalid::placeholder{color:#b11013}a{-webkit-tap-highlight-color:rgba(0,0,0,0)}a:active{opacity:.8}a.wrapper{text-decoration:none}@keyframes spin{to{transform:rotate(360deg)}}.animate-spin{animation:spin 1s linear infinite}
